requirements_swift.txt
torch==2.7.1
transformers==4.55.4
deepspeed==0.16.9
peft==0.17.1
trl==0.20.0
ms-swift==3.7.2
tqdm==4.67.1
numpy==2.2.6
jieba==0.42.1
bert-score==0.3.13
nltk==3.9.1
Levenshtein==0.27.3
chinese2digits==6.0.2
cn2an==0.5.23
regex==2025.7.34
pynvml==13.580.65
